feat(web-app): restructure job components and enhance job status display
- Moved JobsTable and JobStatusBadge components to a shared components directory for better organization.
- Updated imports in the relevant files to reflect the new component paths.
- Introduced a new JobStatusBadge component to display job statuses with appropriate styling in the JobsTable.
