+++ web-app/src/app/app/jobs/components/jobs-table.tsx
@@ -4,7 +4,7 @@ import { ColumnDef, createColumnHelper } from "@tanstack/react-table";
 import { useFormatter, useTranslations } from "next-intl";
 import React from "react";
 
-import JobStatusBadge from "@/app/jobs/job-status-badge";
+import JobStatusBadge from "@/app/jobs/components/job-status-badge";
 import { DataTable, DataTableColumnHeader } from "@/components/data-table";
 import { JobWithRelations } from "@/lib/db/services/job.service";
